feat(yadm): set linuxbrew path in bootstrap
Linuxbrew is a very useful package manager for Linux users,
espicailly for those distributions which do not have newer version of packages.
Besides, it is also useful for those who do not have root access to the system,
since it can be installed in any directory
(it may need to short than the default path `/home/linuxbrew/.linuxbrew`).
For custom installation, the path to linuxbrew should be set manually.