feat(hammerspoon): hotkey window for kitty
Hammerspoon is a powerful automation tool for macOS.
I use it to manage my kitty windows, so I can open a kitty window with a hotkey.
But it doesn't work well as iTerm2, so I will use iTerm2 instead of kitty.
So I will not use this feature until I have a new idea.