chore: migrate from `null-ls.nvim` to `none-ls.nvim`
`null-ls.nvim` is now deprecated in favor of `none-ls.nvim`. This
commit migrates to the new plugin.

The new plugin is a fork of the old one, with the only difference
being the name. The new plugin is also published to the official
plugin registry, so it can be installed with `Packer`.

The new plugin is also a drop-in replacement for the old one, so
no changes are required in the configuration.