feat(vim): use transparent_background
By default, the background color of floating windows is not the same as
the main window.
With this change, the background color of floating windows will be removed,
which will make the floating windows look the same as the main window.

Besides, the winblend and pumblend options are removed to make sure the
transparency works correctly.